+++ ntex-tls/CHANGES.md
@@ -1,5 +1,11 @@
 # Changes
 
+## [0.1.0-b.3] - 2021-12-23
+
+* Add impl openssl::Acceptor::from(SslAcceptor)
+
+* Add openssl::Acceptor::seal() helper
+
 ## [0.1.0-b.1] - 2021-12-20
 
 * Update ntex-io